fix ETL mapping validation TypeError
